+++ src/main/java/com/example/project_3/service/PostService.java
@@ -7,15 +7,13 @@ import com.example.project_3.entity.Post;
 import com.example.project_3.entity.Tag;
 import com.example.project_3.entity.User;
 import com.example.project_3.exception.PostNotFoundException;
-import com.example.project_3.exception.UserNotFoundException;
 import com.example.project_3.mapper.PostMapper;
 import com.example.project_3.repository.PostRepository;
 import com.example.project_3.repository.TagRepository;
-import com.example.project_3.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,47 +21,52 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
+@Transactional(readOnly = true)
 public class PostService {
     private final PostRepository postRepository;
     private final PostMapper postMapper;
-    private final UserRepository userRepository;
     private final TagRepository tagRepository;
 
 
     //Create
-    public PostResponseDTO createPost(PostRequestDTO dto, Long userId){
-        User user = userRepository.findById(userId).
-                orElseThrow(() -> new UserNotFoundException("Пользователь с таким id не найден!"));
+    @Transactional
+    public PostResponseDTO createPost(PostRequestDTO dto, User currentUser){
         Post post = postMapper.toEntity(dto);
-        post.setUser(user);
-
-        if(!dto.getTagIds().isEmpty()){
-            List<Tag> tags = tagRepository.findAllById(dto.getTagIds());
-            post.setTagList(tags);
+        post.setUser(currentUser);
+        if(!dto.tagIds().isEmpty()){
+            List<Tag> tagsList = tagRepository.findAllById(dto.tagIds());
+            post.setTagList(tagsList);
         }
-
         return postMapper.toResponseDTO(postRepository.save(post));
     }
 
     //update
-    public PostResponseDTO updatePostById(PostUpdateDTO dto, Long id){
+    @Transactional
+    public PostResponseDTO updatePostById(PostUpdateDTO dto, Long id, User currentUser){
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Пост с таким id не был найден"));
+
+        if(!post.getUser().getId().equals(currentUser.getId())){
+            throw new AccessDeniedException("Нет доступа!");
+        }
+
         postMapper.update(dto, post);
         return postMapper.toResponseDTO(postRepository.save(post));
     }
 
     //Delete
-    public ResponseEntity<Void> deletePostById(Long id){
+    @Transactional
+    public void deletePostById(Long id, User currentUser){
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Пост с таким id не был найден"));
+        if(!post.getUser().getId().equals(currentUser.getId())){
+            throw new AccessDeniedException("Нет доступа!");
+        }
         postRepository.delete(post);
-        return ResponseEntity.noContent().build();
+
     }
 
     //Read
-    @Transactional(readOnly = true)
     public PostResponseDTO findPostById(Long id){
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Пост с таким id не был найден"));
@@ -71,19 +74,13 @@ public class PostService {
         return postMapper.toResponseDTO(post);
     }
 
-    @Transactional(readOnly = true)
     public Page<PostResponseDTO>  findPostsByTagId(Long tagId, Pageable pageable){
         return postRepository.findPostsByTagListId(tagId, pageable)
                 .map(postMapper::toResponseDTO);
     }
 
-    @Transactional(readOnly = true)
+
     public Page<PostResponseDTO> findPostsByUserId(Long userId, Pageable pageable){
-
-        StringBuilder s = new StringBuilder("ssdf");
-        s.deleteCharAt(1);
-
-
         return postRepository.findPostsByUserId(userId, pageable)
                 .map(postMapper::toResponseDTO);
     }
